+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - Add ability to start game from world selection screen. Can be started in combined and dedicated server mode.
 
 ### Changed
+- Allow committing compressed layers.
+- Convert chunk storage to Mallocator.
 - Free chunk meshes immediately after upload. Saves huge amount of memory.
 - Reuse buffers for meshing (Significantly lowers memory consumption).
 - Optimize getting snapshots for meshing.